# services/synthetic-phi-service/openapi.yaml
openapi: 3.0.3
info:
  title: Synthetic PHI Service
  description: |
    Safe synthetic patient data generation for healthcare compliance testing.
    
    **HIPAA Compliance**: This service generates only synthetic data. No real PHI is exposed.
    
    **Features**:
    - Generate single or batch synthetic patient records
    - HIPAA-compliant data structures
    - Tamper-evident audit trails
    - SOX-compliant request tracking
    
    **Compliance Frameworks**:
    - HIPAA (Health Insurance Portability and Accountability Act)
    - FDA 21 CFR Part 11 (Electronic Records)
    - SOX (Sarbanes-Oxley Act)
  version: 1.0.0
  contact:
    name: GitOps 2.0 Team
    url: https://github.com/ITcredibl/gitops2-enterprise-git-intel-demo
  license:
    name: MIT

servers:
  - url: http://localhost:8081
    description: Local development server
  - url: https://synthetic-phi.example.com
    description: Production server

tags:
  - name: Patient Data
    description: Synthetic patient record generation
  - name: Health
    description: Service health and readiness checks
  - name: Compliance
    description: Compliance status and validation

paths:
  /synthetic-patient:
    get:
      tags:
        - Patient Data
      summary: Generate single synthetic patient
      description: |
        Generates a single synthetic patient record with randomized but safe data.
        All data is synthetic and HIPAA-compliant for testing purposes.
      operationId: generateSinglePatient
      responses:
        '200':
          description: Successfully generated synthetic patient
          headers:
            X-PHI-Protected:
              description: Indicates PHI protection status
              schema:
                type: string
                enum: [true]
            X-Audit-Trail:
              description: Audit trail status
              schema:
                type: string
                enum: [enabled]
            X-Encryption-Status:
              description: Encryption algorithm used
              schema:
                type: string
                example: AES-256
            X-SOX-Compliant:
              description: SOX compliance indicator
              schema:
                type: string
                enum: [true]
            X-Request-ID:
              description: Unique request identifier for audit trail
              schema:
                type: string
                example: REQ_1234567890
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyntheticPatient'
              example:
                id: PAT123456
                first_name: Jordan
                last_name: Smith
                date_of_birth: "1975-01-01"
                medical_record_number: MRN789012
                diagnosis:
                  - Hypertension
                  - Diabetes Type 2
                created_at: "2025-11-23T10:30:00Z"
                encrypted: true
                phi_tags:
                  - name
                  - dob
                  - mrn
                  - diagnosis
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Patient Data
      summary: Generate batch of synthetic patients
      description: |
        Generates multiple synthetic patient records (up to 100 per request).
        Useful for populating test databases or load testing.
      operationId: generateBatchPatients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - count
              properties:
                count:
                  type: integer
                  minimum: 1
                  maximum: 100
                  description: Number of synthetic patients to generate
                  example: 10
      responses:
        '200':
          description: Successfully generated batch of synthetic patients
          headers:
            X-PHI-Protected:
              description: Indicates PHI protection status
              schema:
                type: string
                enum: [true]
            X-Audit-Trail:
              description: Audit trail status
              schema:
                type: string
                enum: [enabled]
            X-Request-ID:
              description: Unique request identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  patients:
                    type: array
                    items:
                      $ref: '#/components/schemas/SyntheticPatient'
                  count:
                    type: integer
                    description: Number of patients generated
                  metadata:
                    type: object
                    properties:
                      generation_time:
                        type: string
                        format: date-time
                      compliance:
                        type: string
                        example: HIPAA_SYNTHETIC
                      phi_safe:
                        type: string
                        example: "true"
        '400':
          description: Invalid request (count out of range or malformed JSON)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /health:
    get:
      tags:
        - Health
      summary: Health check endpoint
      description: |
        Returns the service health status. Use for Kubernetes liveness probes.
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [healthy, degraded, unhealthy]
                  service:
                    type: string
                    example: synthetic-phi-service
                  version:
                    type: string
                    example: 1.0.0
                  timestamp:
                    type: string
                    format: date-time
                  compliance:
                    type: array
                    items:
                      type: string
                    example: [HIPAA, synthetic_only]

  /compliance/status:
    get:
      tags:
        - Compliance
      summary: Compliance status endpoint
      description: |
        Returns detailed compliance status for HIPAA, SOX, and privacy frameworks.
        Use for compliance audits and regulatory validation.
      operationId: complianceStatus
      responses:
        '200':
          description: Compliance status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  hipaa_compliant:
                    type: boolean
                    description: HIPAA compliance status
                  phi_data_synthetic:
                    type: boolean
                    description: Whether data is synthetic (true) or real (false)
                  real_phi_exposure:
                    type: boolean
                    description: Whether real PHI is exposed
                  encryption_enabled:
                    type: boolean
                    description: Data encryption status
                  audit_trail_active:
                    type: boolean
                    description: Audit logging status
                  compliance_verified:
                    type: string
                    format: date-time
                    description: Last compliance verification timestamp
                  privacy_by_design:
                    type: boolean
                    description: Privacy-by-design implementation
                  data_minimization:
                    type: boolean
                    description: Data minimization principle compliance
                  purpose_limitation:
                    type: string
                    description: Purpose limitation statement
                    example: testing_only

  /metrics:
    get:
      tags:
        - Health
      summary: Prometheus metrics endpoint
      description: |
        Returns Prometheus-formatted metrics for monitoring.
        
        **Metrics exposed**:
        - `synthetic_phi_patients_generated_total` - Total patients generated
        - `synthetic_phi_request_duration_seconds` - Request duration histogram
        - `synthetic_phi_active_requests` - Active request gauge
        - `synthetic_phi_compliance_checks_total` - Compliance check counter
      operationId: prometheusMetrics
      responses:
        '200':
          description: Prometheus metrics
          content:
            text/plain:
              schema:
                type: string
              example: |
                # HELP synthetic_phi_patients_generated_total Total number of synthetic patients generated
                # TYPE synthetic_phi_patients_generated_total counter
                synthetic_phi_patients_generated_total 1234

  /:
    get:
      tags:
        - Health
      summary: Service information endpoint
      description: Returns basic service information and available endpoints
      operationId: serviceInfo
      responses:
        '200':
          description: Service information
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    type: string
                  description:
                    type: string
                  version:
                    type: string
                  endpoints:
                    type: object
                  compliance:
                    type: object

components:
  schemas:
    SyntheticPatient:
      type: object
      description: Synthetic patient record (HIPAA-safe)
      required:
        - id
        - first_name
        - last_name
        - date_of_birth
        - medical_record_number
        - diagnosis
        - created_at
        - encrypted
        - phi_tags
      properties:
        id:
          type: string
          description: Synthetic patient identifier
          pattern: '^PAT[0-9]{6}$'
          example: PAT123456
        first_name:
          type: string
          description: Synthetic first name (from safe list)
          example: Jordan
        last_name:
          type: string
          description: Synthetic last name (from safe list)
          example: Smith
        date_of_birth:
          type: string
          format: date
          description: Synthetic date of birth (1950-2000)
          example: "1975-01-01"
        medical_record_number:
          type: string
          description: Synthetic medical record number
          pattern: '^MRN[0-9]{6}$'
          example: MRN789012
        diagnosis:
          type: array
          description: List of synthetic diagnoses (1-3)
          minItems: 1
          maxItems: 3
          items:
            type: string
          example:
            - Hypertension
            - Diabetes Type 2
        created_at:
          type: string
          format: date-time
          description: Record creation timestamp
        encrypted:
          type: boolean
          description: Encryption status flag
          example: true
        phi_tags:
          type: array
          description: PHI element tags
          items:
            type: string
          example:
            - name
            - dob
            - mrn
            - diagnosis
    
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Invalid request body

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future use)

security:
  - {}  # No authentication required for synthetic data generation